+++ runners/twister2/src/main/java/org/apache/beam/runners/twister2/translators/functions/ImpulseSource.java
@@ -1,0 +1,43 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.beam.runners.twister2.translators.functions;
+
+import edu.iu.dsc.tws.api.tset.fn.SourceFunc;
+import org.apache.beam.sdk.util.WindowedValue;
+
+/** A {@link SourceFunc} which executes the impulse transform contract. */
+public class ImpulseSource implements SourceFunc<WindowedValue<byte[]>> {
+  private static final WindowedValue<byte[]> IMPULSE_VALUE =
+      WindowedValue.valueInGlobalWindow(new byte[0]);
+
+  private boolean impulseEmitted = false;
+
+  @Override
+  public boolean hasNext() {
+    return !impulseEmitted;
+  }
+
+  @Override
+  public WindowedValue<byte[]> next() {
+    if (impulseEmitted) {
+      throw new IllegalStateException("Impulse has already emitted an element.");
+    }
+    impulseEmitted = true;
+    return IMPULSE_VALUE;
+  }
+}
